data/requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorclass==2.2.2
docopt==0.6.2
geojson==2.5.0
idna==3.3
packaging==21.3
pip-upgrader==1.4.15
psycopg2-binary==2.9.3
pyparsing==3.0.9
PyYAML==6.0
requests==2.27.1
Shapely==1.8.2
terminaltables==3.1.10
urllib3==1.26.9
xmltodict==0.13.0
